Update settings config in test runner config
